source/mre-fan-experience-frontend/cdk/requirements.txt
aws-cdk.aws_iam
aws_cdk.aws_codecommit
aws_cdk.aws_amplify
aws_cdk.aws_cognito
boto3